moved App from src to src/Components/App
